Consolidate class extensions
 * Add `Class<*>.urlClassLoader()` extension.
 * Rename `Fixtures` to `BuildResultExts`.
